Check plugin_extras type
Added explicit checks that the plugin_extras field in the user object is read as a Python dict, and only then retrieving the saml2auth dict and the saml_id key value. This fixes an error where existing user records in the database had plugin_extras values that were not as expected, and thus did not get read as a dict.